+++ pai-dlc-20201203/ChangeLog.md
@@ -1,3 +1,8 @@
+2024-06-06 Version: 1.4.6
+- Update API ListTensorboards: add param UserId.
+- Update API ListTensorboards: add param Username.
+
+
 2024-04-09 Version: 1.4.5
 - Update API CreateTensorboard: update param body.
 - Update API ListJobs: add param ResourceQuotaName.